+++ src/libs/logs/FactoryLogs.sol
@@ -3,6 +3,6 @@ pragma solidity 0.8.30;
 
 import {ExtraTicketData} from "@host-it-storage/FactoryStorage.sol";
 
-event TicketCreated(uint256 indexed ticketId, address indexed ticketAdmin, ExtraTicketData ticketData);
+event TicketCreated(uint56 indexed ticketId, address indexed ticketAdmin, ExtraTicketData ticketData);
 
-event TicketUpdated(uint256 indexed ticketId, address indexed ticketAdmin, ExtraTicketData ticketData);
+event TicketUpdated(uint56 indexed ticketId, address indexed ticketAdmin, ExtraTicketData ticketData);
